+++ src/SysadminsLV.PKI/Cryptography/TspStatusInfo.cs
@@ -30,7 +30,7 @@ public class TspStatusInfo {
     public String ErrorText { get; private set; }
 
     void decode(Asn1Reader asn) {
-        asn.MoveNextAndExpectTags((Byte)Asn1Type.INTEGER);
+        asn.MoveNextAndExpectTags(Asn1Type.INTEGER);
         ResponseStatus = (TspResponseStatus)(Int32)new Asn1Integer(asn).Value;
         while (asn.MoveNextSibling()) {
             switch (asn.Tag) {
