+++ src/SysadminsLV.PKI/Cryptography/X509Certificates/X509CertificateTemplateExtension.cs
@@ -75,11 +75,11 @@ public sealed class X509CertificateTemplateExtension : X509Extension {
     }
     void m_decode(Byte[] rawData) {
         var asn = new Asn1Reader(rawData);
-        asn.MoveNextSiblingAndExpectTags((Byte)Asn1Type.OBJECT_IDENTIFIER);
+        asn.MoveNextSiblingAndExpectTags(Asn1Type.OBJECT_IDENTIFIER);
         TemplateOid = new Asn1ObjectIdentifier(asn.GetTagRawData()).Value;
-        asn.MoveNextAndExpectTags((Byte)Asn1Type.INTEGER);
+        asn.MoveNextAndExpectTags(Asn1Type.INTEGER);
         MajorVersion = (Int32)new Asn1Integer(asn.GetTagRawData()).Value;
-        asn.MoveNextAndExpectTags((Byte)Asn1Type.INTEGER);
+        asn.MoveNextAndExpectTags(Asn1Type.INTEGER);
         MinorVersion = (Int32)new Asn1Integer(asn.GetTagRawData()).Value;
         RawData = rawData.ToArray();
     }
